+++ src/app/game/game.module.ts
@@ -72,6 +72,7 @@ import { MinimapComponent } from './map/map-canvas/minimap/minimap.component';
 import { TutorialComponent } from './tutorial/tutorial.component';
 import { TutorialDialogComponent } from './tutorial/tutorial-dialog/tutorial-dialog.component';
 import { AllianceInvitationsComponent } from './alliance/alliance-invitations/alliance-invitations.component';
+
 export const effects = [
   PlayerEffects,
   Allianceffects,
